Improved signifiers for video extraction
See title. Added a prompt to extract and made it clear that video footage is cleared from the camera after extraction is complete.

Additionally, the player can now only extract the footage at a certain distance away from the terminal, to prevent doing it on accident and out of view.

Time taken: 40 mins.